Check for test data relative to running executable (#165)
* Check for test data relative to running executable

* Format check
