+++ clients/common/utility.cpp
@@ -1,5 +1,5 @@
 /* ************************************************************************
- * Copyright (C) 2020-2022 Advanced Micro Devices, Inc. All rights reserved.
+ * Copyright (C) 2020-2023 Advanced Micro Devices, Inc. All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -107,65 +107,6 @@ double get_time_us_sync(hipStream_t stream)
 {
     hipStreamSynchronize(stream);
     return get_time_us_no_sync();
-}
-
-/* ============================================================================================ */
-/*  device query and print out their ID and name; return number of compute-capable devices. */
-int query_device_property()
-{
-    int               device_count;
-    hipsolverStatus_t count_status = (hipsolverStatus_t)hipGetDeviceCount(&device_count);
-    if(count_status != HIPSOLVER_STATUS_SUCCESS)
-    {
-        printf("Query device error: cannot get device count \n");
-        return -1;
-    }
-    else
-    {
-        printf("Query device success: there are %d devices \n", device_count);
-    }
-
-    for(int i = 0; i < device_count; i++)
-    {
-        hipDeviceProp_t   props;
-        hipsolverStatus_t props_status = (hipsolverStatus_t)hipGetDeviceProperties(&props, i);
-        if(props_status != HIPSOLVER_STATUS_SUCCESS)
-        {
-            printf("Query device error: cannot get device ID %d's property\n", i);
-        }
-        else
-        {
-            printf("Device ID %d : %s ------------------------------------------------------\n",
-                   i,
-                   props.name);
-            printf("with %3.1f GB memory, clock rate %dMHz @ computing capability %d.%d \n",
-                   props.totalGlobalMem / 1e9,
-                   (int)(props.clockRate / 1000),
-                   props.major,
-                   props.minor);
-            printf(
-                "maxGridDimX %d, sharedMemPerBlock %3.1f KB, maxThreadsPerBlock %d, warpSize %d\n",
-                props.maxGridSize[0],
-                props.sharedMemPerBlock / 1e3,
-                props.maxThreadsPerBlock,
-                props.warpSize);
-
-            printf("-------------------------------------------------------------------------\n");
-        }
-    }
-
-    return device_count;
-}
-
-/*  set current device to device_id */
-void set_device(int device_id)
-{
-    hipsolverStatus_t status = (hipsolverStatus_t)hipSetDevice(device_id);
-    if(status != HIPSOLVER_STATUS_SUCCESS)
-    {
-        printf("Set device error: cannot set device ID %d, there may not be such device ID\n",
-               (int)device_id);
-    }
 }
 
 #ifdef __cplusplus
