+++ clients/include/clientcommon.hpp
@@ -1,5 +1,5 @@
 /* ************************************************************************
- * Copyright (C) 2020-2022 Advanced Micro Devices, Inc. All rights reserved.
+ * Copyright (C) 2020-2023 Advanced Micro Devices, Inc. All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,6 +23,7 @@
 
 #pragma once
 
+#include "../rocblascommon/clients_utility.hpp"
 #include "../rocblascommon/rocblas_vector.hpp"
 #include "../rocsolvercommon/norm.hpp"
 #include "../rocsolvercommon/rocsolver_arguments.hpp"
